+++ packages/sdk/test/unit/sdkBase.spec.ts
@@ -8,12 +8,14 @@ import {
   AuctionBid,
   Logger,
   delay,
+  GAS_ESTIMATES,
 } from "@connext/nxtp-utils";
 import { expect } from "chai";
 import { providers, Wallet, constants, BigNumber } from "ethers";
 import { createStubInstance, reset, restore, SinonStub, SinonStubbedInstance, stub } from "sinon";
 
 import { MAX_SLIPPAGE_TOLERANCE, MIN_SLIPPAGE_TOLERANCE } from "../../src/sdk";
+import * as TransactionManagerHelperFns from "../../src/transactionManager/transactionManager";
 
 import * as utils from "../../src/utils";
 import * as sdkIndex from "../../src/sdkBase";
@@ -32,6 +34,7 @@ import {
   SubgraphsNotSynced,
   UnknownAuctionError,
   InvalidCallTo,
+  NoValidBids,
 } from "../../src/error";
 import { getAddress, keccak256 } from "ethers/lib/utils";
 import { CrossChainParams, NxtpSdkEvents, HistoricalTransactionStatus } from "../../src";
@@ -55,6 +58,8 @@ describe("NxtpSdkBase", () => {
   let messaging: SinonStubbedInstance<UserNxtpNatsMessagingService>;
   let subgraph: SinonStubbedInstance<Subgraph>;
   let transactionManager: SinonStubbedInstance<TransactionManager>;
+  let getDeployedChainIdsForGasFeeStub;
+  let getDeployedPriceOracleContractStub;
   let provider1337: SinonStubbedInstance<providers.FallbackProvider>;
   let provider1338: SinonStubbedInstance<providers.FallbackProvider>;
   let signFulfillTransactionPayloadMock: SinonStub;
@@ -96,7 +101,10 @@ describe("NxtpSdkBase", () => {
     transactionManager = createStubInstance(TransactionManager);
 
     stub(utils, "getDecimals").resolves(18);
+    stub(utils, "getTokenPrice").resolves(BigNumber.from(10).pow(21));
     stub(utils, "getTimestampInSeconds").resolves(Math.floor(Date.now() / 1000));
+    provider1337.getGasPrice.resolves(BigNumber.from(10).pow(9));
+    provider1338.getGasPrice.resolves(BigNumber.from(10).pow(9));
 
     balanceStub = stub(utils, "getOnchainBalance");
     balanceStub.resolves(BigNumber.from(0));
@@ -130,6 +138,16 @@ describe("NxtpSdkBase", () => {
     (sdk as any).messaging = messaging;
 
     messaging.connect.resolves(response);
+
+    getDeployedChainIdsForGasFeeStub = stub(TransactionManagerHelperFns, "getDeployedChainIdsForGasFee").returns([
+      sendingChainId,
+      receivingChainId,
+    ]);
+
+    getDeployedPriceOracleContractStub = stub(TransactionManagerHelperFns, "getDeployedPriceOracleContract").returns({
+      address: "0x0",
+      abi: "",
+    });
   });
 
   afterEach(() => {
@@ -175,11 +193,13 @@ describe("NxtpSdkBase", () => {
     auctionBidOverrides: Partial<AuctionBid> = {},
     _bidSignature: string = EmptyCallDataHash,
     _gasFeeInReceivingToken = "0",
+    _metaTxRelayerFe = "0",
   ): {
     crossChainParams: CrossChainParams;
     auctionBid: AuctionBid;
     bidSignature: string;
     gasFeeInReceivingToken: string;
+    metaTxRelayerFee: string;
   } => {
     const transactionId = getRandomBytes32();
     const crossChainParams = {
@@ -221,8 +241,9 @@ describe("NxtpSdkBase", () => {
 
     const bidSignature = _bidSignature;
     const gasFeeInReceivingToken = _gasFeeInReceivingToken;
+    const metaTxRelayerFee = _metaTxRelayerFe;
 
-    return { crossChainParams, auctionBid, bidSignature, gasFeeInReceivingToken };
+    return { crossChainParams, auctionBid, bidSignature, gasFeeInReceivingToken, metaTxRelayerFee };
   };
 
   describe("#constructor", () => {
@@ -248,6 +269,7 @@ describe("NxtpSdkBase", () => {
     });
 
     it("should error if subgraph doesn't exist for chainId", async () => {
+      getDeployedChainIdsForGasFeeStub.returns([sendingChainId, receivingChainId]);
       const _chainConfig = {
         [sendingChainId]: {
           provider: provider1337,
@@ -430,7 +452,7 @@ describe("NxtpSdkBase", () => {
         messageEvt.post({ inbox: "inbox", data: { bidSignature, bid: auctionBid } });
       }, 200);
       await expect(sdk.getTransferQuote(crossChainParams)).to.eventually.be.rejectedWith(
-        UnknownAuctionError.getMessage(auctionBid.transactionId),
+        NoValidBids.getMessage(auctionBid.transactionId),
       );
     });
 
@@ -445,7 +467,7 @@ describe("NxtpSdkBase", () => {
         messageEvt.post({ inbox: "inbox", data: { bidSignature, bid: auctionBid } });
       }, 200);
       await expect(sdk.getTransferQuote(crossChainParams)).to.eventually.be.rejectedWith(
-        UnknownAuctionError.getMessage(auctionBid.transactionId),
+        NoValidBids.getMessage(auctionBid.transactionId),
       );
     });
 
@@ -460,7 +482,7 @@ describe("NxtpSdkBase", () => {
         messageEvt.post({ inbox: "inbox", data: { bidSignature, bid: auctionBid } });
       }, 200);
       await expect(sdk.getTransferQuote(crossChainParams)).to.eventually.be.rejectedWith(
-        UnknownAuctionError.getMessage(auctionBid.transactionId),
+        NoValidBids.getMessage(auctionBid.transactionId),
       );
     });
 
@@ -636,7 +658,11 @@ describe("NxtpSdkBase", () => {
 
       transactionManager.prepare.resolves(TxRequest);
 
-      const res = await sdk.prepareTransfer({ bid: auctionBid, bidSignature, gasFeeInReceivingToken });
+      const res = await sdk.prepareTransfer({
+        bid: auctionBid,
+        bidSignature,
+        gasFeeInReceivingToken,
+      });
       expect(res).to.deep.eq(TxRequest);
     });
   });
@@ -899,6 +925,154 @@ describe("NxtpSdkBase", () => {
       );
 
       expect(res).to.deep.eq(TxRequest);
+    });
+  });
+
+  describe("#estimateHardcodedFeeForPrepare", () => {
+    it("happy: should return valid price if price oracle is configured", async () => {
+      const { crossChainParams } = getMock();
+
+      const result = await sdk.estimateHardcodedFeeForPrepare(
+        crossChainParams.sendingChainId,
+        crossChainParams.sendingAssetId,
+        18,
+        null,
+        null,
+      );
+
+      expect(result).to.be.eq(BigNumber.from(GAS_ESTIMATES.prepare).mul("1000000000"));
+    });
+
+    it("happy: should return zero price if price oracle isn't configured", async () => {
+      getDeployedPriceOracleContractStub.returns({ address: null, abi: null });
+      const { crossChainParams } = getMock();
+
+      const result = await sdk.estimateHardcodedFeeForPrepare(
+        crossChainParams.sendingChainId,
+        crossChainParams.sendingAssetId,
+        18,
+        null,
+        null,
+      );
+
+      expect(result).to.be.eq(BigNumber.from("0"));
+    });
+  });
+
+  describe("#estimateHardcodedFeeForFulfill", () => {
+    it("happy: should return valid price if price oracle is configured", async () => {
+      const { crossChainParams } = getMock();
+
+      const result = await sdk.estimateHardcodedFeeForFulfill(
+        crossChainParams.sendingChainId,
+        crossChainParams.sendingAssetId,
+        18,
+        null,
+        null,
+      );
+
+      expect(result).to.be.eq(BigNumber.from(GAS_ESTIMATES.fulfill).mul("1000000000"));
+    });
+    it("happy: should return zero price if price oracle isn't configured", async () => {
+      getDeployedPriceOracleContractStub.returns({ address: null, abi: null });
+      const { crossChainParams } = getMock();
+
+      const result = await sdk.estimateHardcodedFeeForFulfill(
+        crossChainParams.sendingChainId,
+        crossChainParams.sendingAssetId,
+        18,
+        null,
+        null,
+      );
+
+      expect(result).to.be.eq(BigNumber.from("0"));
+    });
+  });
+
+  describe("#estimateFeeForRouterTransfer", () => {
+    it("happy: should return valid price if price oracle is configured", async () => {
+      const { crossChainParams } = getMock();
+      const result = await sdk.estimateFeeForRouterTransfer(
+        crossChainParams.sendingChainId,
+        crossChainParams.sendingAssetId,
+        crossChainParams.receivingChainId,
+        crossChainParams.receivingAssetId,
+        true,
+        null,
+        null,
+      );
+
+      expect(result).to.be.eq(
+        BigNumber.from(GAS_ESTIMATES.fulfill).add(BigNumber.from(GAS_ESTIMATES.prepare)).mul("1000000000"),
+      );
+    });
+    it("happy: should return zero price if price oracle isn't configured", async () => {
+      const { crossChainParams } = getMock();
+      getDeployedChainIdsForGasFeeStub.returns([11111, 22222]);
+      const result = await sdk.estimateFeeForRouterTransfer(
+        crossChainParams.sendingChainId,
+        crossChainParams.sendingAssetId,
+        crossChainParams.receivingChainId,
+        crossChainParams.receivingAssetId,
+        true,
+        null,
+        null,
+      );
+      expect(result).to.be.eq("0");
+    });
+    it("should error for non configured chains", async () => {
+      await expect(
+        sdk.estimateFeeForRouterTransfer(111, "0x0", 222, "0x0", true, null, null),
+      ).to.eventually.rejectedWith(ChainNotConfigured.getMessage(111, supportedChains));
+    });
+  });
+
+  describe("#estimateFeeForMetaTx", () => {
+    it("happy: should return valid price if price oracle is configured", async () => {
+      const { crossChainParams } = getMock();
+      const result = await sdk.estimateFeeForMetaTx(
+        crossChainParams.sendingChainId,
+        crossChainParams.sendingAssetId,
+        crossChainParams.receivingChainId,
+        crossChainParams.receivingAssetId,
+        true,
+        null,
+        null,
+      );
+
+      expect(result).to.be.eq(BigNumber.from(GAS_ESTIMATES.fulfill).mul("1000000000"));
+    });
+    it("happy: should return zero price if price oracle isn't configured", async () => {
+      const { crossChainParams } = getMock();
+      getDeployedChainIdsForGasFeeStub.returns([11111, 22222]);
+      const result = await sdk.estimateFeeForMetaTx(
+        crossChainParams.sendingChainId,
+        crossChainParams.sendingAssetId,
+        crossChainParams.receivingChainId,
+        crossChainParams.receivingAssetId,
+        true,
+        null,
+        null,
+      );
+      expect(result).to.be.eq("0");
+    });
+    it("should error for non configured chains", async () => {
+      await expect(sdk.estimateFeeForMetaTx(111, "0x0", 222, "0x0", true, null, null)).to.eventually.rejectedWith(
+        ChainNotConfigured.getMessage(111, supportedChains),
+      );
+    });
+  });
+
+  describe("#getGasPrice", () => {
+    it("should work for configured chain", async () => {
+      const result = await sdk.getGasPrice(sendingChainId);
+      expect(result).to.be.eq(BigNumber.from("1000000000"));
+    });
+
+    it("should error for non-configured chain", async () => {
+      await expect(sdk.getGasPrice(11111)).eventually.be.rejectedWith(
+        ChainNotConfigured.getMessage(11111, supportedChains),
+      );
     });
   });