+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Next Release
 
+## 0.0.70
+
+- [router] Handle fulfill relay
+- [router/sdk] Add subgraph fallbacks by default
+- [test-ui] Fixes for test-ui based on feedback
+
+## 0.0.69
+
+- [sdk] Return fulfill fee from `getTransferQuote`
+
 ## 0.0.68
 
 - [sdk] Fix build